Added optional parameter called "afterModalOpen"
Added optional parameter called "afterModalOpen" that accepts a function that gets called after the modal opens. This allows the user to interact with the DOM in the modal.

Example: I have a popup modal that has one input that accepts an email. But under the input is a button that you can press to append a new input if you want to send the email to more people. This is not possible without the "afterModalOpen" parameter.