feat(api): inputValue as a Promise (#1025)
* New Feature: handle inputValue as promise
Implement handling of inputValue as promise similar
to what is already inplmented for select and radio.

Before showing the modal, if the inputValue is a promise,
then the input is hidden wating for the promise to be
resolved  / rejected. Once the promise is resolved, the result
is assigned to the value of the input. If the promise is
rejected, the input value is set to empty string and an error
is reported

* Make list of applicable input type explicit

* Update Typescript definition of inputValue to include Promise

* Handling parsing of number in case input type is a number

* Added test case for input value as promise