fix(api): falsy values in preConfirm (#1403)
* Changed check on preConfirmValue

Currently the check on preConfirmValue is done on a falsy value. This include also 0,
which could be a valid return value.

Replaced it with a check for udnefined value, which is the only case
where we may want to return something else

* Added test for preConfirm returning undefined

* Changed test for preConfirm return value

Testing the code that causes the issue this time.